refactor(Dialog): refactor `Dialog` components
