+++ src/components/dialog/dialog-description.tsx
@@ -1,10 +1,11 @@
 import * as React from "react";
 
 import { Dialog as ReactDialog } from "radix-ui";
-import { twMerge } from "tailwind-merge";
+
+import { cn } from "~/utils";
 
 export type DialogDescriptionProps = React.ComponentProps<typeof ReactDialog.Description>;
 
 export function DialogDescription({ className, ...props }: DialogDescriptionProps) {
-  return <ReactDialog.Description className={twMerge("text-body-2 text-gray-200", className)} {...props} />;
+  return <ReactDialog.Description className={cn("text-body-2 text-gray-300", className)} {...props} />;
 }
