+++ src/components/dialog/dialog-footer.tsx
@@ -1,11 +1,9 @@
 import * as React from "react";
 
-import { twMerge } from "tailwind-merge";
+import { cn } from "~/utils";
 
 export type DialogFooterProps = React.DetailedHTMLProps<React.HTMLAttributes<HTMLDivElement>, HTMLDivElement>;
 
 export function DialogFooter({ className, ...props }: DialogFooterProps) {
-  return (
-    <div className={twMerge("mt-4 flex flex-col-reverse gap-2 sm:flex-row sm:justify-end", className)} {...props} />
-  );
+  return <div className={cn("mt-4 flex flex-col-reverse gap-2 sm:flex-row sm:justify-end", className)} {...props} />;
 }
