+++ src/components/dialog/dialog-title.tsx
@@ -1,10 +1,11 @@
 import * as React from "react";
 
 import { Dialog as ReactDialog } from "radix-ui";
-import { twMerge } from "tailwind-merge";
+
+import { cn } from "~/utils";
 
 export type DialogTitleProps = React.ComponentProps<typeof ReactDialog.Title>;
 
 export function DialogTitle({ className, ...props }: DialogTitleProps) {
-  return <ReactDialog.Title className={twMerge("text-heading-6", className)} {...props} />;
+  return <ReactDialog.Title className={cn("text-heading-6", className)} {...props} />;
 }
